+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Version 3.316
 Automatic install detection for the first installation
 code refactoring
 add support menu with wiki link (technical doc)
-
+error template updated
+librairies updated
 
 
 -------------