+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Automatic install detection for the first installation
 code refactoring
 add support menu with wiki link (technical doc)
 error template updated
-librairies updated
+librairies updated.
 
 
 -------------